% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distcomp.R
\name{availableDataSources}
\alias{availableDataSources}
\title{Return currently implemented data sources}
\usage{
availableDataSources()
}
\value{
a list of named arguments, each of which is another list, with
required fields named \code{desc}, a textual description and
\code{requiredPackages}
}
\description{
The function \code{availableDataSources} returns the
currently implemented data sources such as CSV files, Redcap etc.
}
\examples{
availableDataSources()
}
