% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distcomp.R
\name{createNCPInstance}
\alias{createNCPInstance}
\title{Given the definition identifier of an object, instantiate and store object in workspace}
\usage{
createNCPInstance(number, defnId, instanceId, pubkey, den,
  data_file_name = NULL)
}
\arguments{
\item{number}{1 or 2 for indentifying NC party}

\item{defnId}{the computation definition id}

\item{instanceId}{an indentifier to use for the created instance}

\item{pubkey}{the public key}

\item{den}{the denominator for rational approximations}

\item{dataFileName}{a file name to use for saving the data. Typically \code{NULL}, this
is only needed when one is using a single opencpu server to behave like multiple
sites in which case the data file name serves to distinguish the site-specific data files.
When it is \code{NULL}, the data file name is taken from the configuration settings}
}
\value{
TRUE if everything goes well
}
\description{
The function \code{createNCPInstance} uses a definition identified by
defnId to create the appropriate object instance. The instantiated object is assigned
the instanceId and saved under the dataFileName if the latter is specified.
This instantiated object may change state between iterations when a computation executes
}
