% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distcomp.R
\name{defineNewComputation}
\alias{defineNewComputation}
\title{Define a new computation}
\usage{
defineNewComputation()
}
\value{
the results of running the web application
}
\description{
This function just calls \code{\link{runDistcompApp}} with the
parameter "definition"
}
\seealso{
\code{\link{runDistcompApp}}
}
