% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/distcomp.R
\name{.makeOpencpuURL}
\alias{.deSerialize}
\alias{.makeOpencpuURL}
\title{Make an appropriate opencpu URL for a specified function and url prefix for the
distcomp package}
\usage{
.makeOpencpuURL(fn, urlPrefix, package = "distcomp")

.deSerialize(q)
}
\arguments{
\item{fn}{is the name of the function in the distcomp package}

\item{urlPrefix}{is the URL of the opencpu server with the distcomp package installed}

\item{q}{the result of a httr response}
}
\value{
the formatted url as a string

the converted result, if JSON, or the raw content
}
\description{
.makeOpencpuURL returns an appropriate URL to call a function in the distcomp
package given the name of the function and a url prefix.

.deSerialize will convert the JSON result of a http response as needed,
else the raw content is returned.
}
\examples{
distcomp:::.makeOpencpuURL("foo", "http://localhost:9999/ocpu")
}
\keyword{internal}

