% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distcomp.R
\name{executeHEMethod}
\alias{executeHEMethod}
\title{Given the id of a serialized object, invoke a method on the object
with arguments using homomorphic encryption}
\usage{
executeHEMethod(objectId, method, ...)
}
\arguments{
\item{objectId}{the (instance) identifier of the object on which to
invoke a method}

\item{method}{the name of the method to invoke}

\item{...}{further arguments as appropriate for the method}
}
\value{
a list containing an integer and a fractional result converted to characters
}
\description{
The function \code{executeHEMethod} is a homomorphic
encryption wrapper around \code{executeMethod}. It ensures any
returned result is encrypted using the homomorphic encryption
function.
}
