% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distcomp.R
\name{executeMethodOnObject}
\alias{executeMethodOnObject}
\title{Given the name of an object in the global environment, invoke a
method on the object with arguments}
\usage{
executeMethodOnObject(objectName, method, ...)
}
\arguments{
\item{objectName}{the identifier of the object that exists in the
global environment}

\item{method}{the name of the method to invoke}

\item{...}{further arguments as appropriate for the method}
}
\value{
a result that depends on the computation being executed
}
\description{
The function \code{executeMethodOnObject} executes an
arbitrary method on an object that has been instantiated in the
global environment.The result, which is dependent on the
computation that is executed, is returned. If the object needs
to save state between iterations on it, it is automatically
serialized back for the ensuing iterations
}
