% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distcomp.R
\name{generateId}
\alias{generateId}
\title{Generate an identifier for an object}
\usage{
generateId(object, algo = "xxhash64")
}
\arguments{
\item{object}{the object for which a hash is desired}

\item{algo}{the algorithm to use, default is "xxhash64" from
\code{\link[digest]{digest}}}
}
\value{
the hash as a string
}
\description{
A hash is generated based on the contents of the object
}
\seealso{
\code{\link[digest]{digest}}
}
