% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distcomp.R
\name{getComputationInfo}
\alias{getComputationInfo}
\title{Get the value of a variable from the global store}
\usage{
getComputationInfo(name)
}
\arguments{
\item{name}{the name for the object}
}
\value{
the value for the variable, \code{NULL} if not set
}
\description{
In distcomp, several web applications need to communicate
between themselves. Since only one application is expected to be
active at any time, they do so via a global store, essentially a hash table.
This function retrieves the value of a name
}
\seealso{
\code{\link[=setComputationInfo]{setComputationInfo()}}
}
