% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master.R
\docType{class}
\name{HEMaster}
\alias{HEMaster}
\title{Create a HEMaster process for use in a distributed homomorphic
computation}
\format{An \code{\link[R6:R6Class]{R6::R6Class()}} generator object}
\usage{
HEMaster
}
\description{
\code{HEMaster} objects run a distributed computation based
upon a definition file that encapsulates all information
necessary to performa computation. A master makes use of two
non-cooperating parties which communicate with sites that
perform the actual computations using local data.
}
\section{Methods}{

\describe{
\item{\code{HEMaster$new(defn)}}{Create a new \code{HEMaster} instance using
definition provided}
\item{\code{addNCP(url, ncpWorker)}}{Add a non-cooperating party using either the \code{url} provided or a \code{ncpWorker} object of class \code{NCP}. The latter is for prototyping locally using created objects and so exactly one of the two has to be non-null}
\item{\code{run()}}{Run a computation}
}
}

\seealso{
\code{\link[=NCP]{NCP()}}
}
\keyword{datasets}
