% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/querycount.R
\docType{class}
\name{HEQueryCountWorker}
\alias{HEQueryCountWorker}
\title{Create a worker object for use as a worker with master objects generated by \code{\link[=HEQueryCountMaster]{HEQueryCountMaster()}}}
\format{An \code{\link[R6:R6Class]{R6::R6Class()}} generator object}
\usage{
HEQueryCountWorker
}
\description{
\code{HEQueryCountWorker} objects are worker objects at each site of
a distributed QueryCount model computation using homomorphic encryption
}
\section{Methods}{


\describe{
\item{\code{HEQueryCountWorker$new(defn, data, pubkey, den)}}{Create a new HEQueryCount worker object with given
definition defn, data, flag for preserving state between iterations, public key and denominator for rational approximations}
\item{\code{setParams(pubkey, den)}}{Set the public key and denominator for rational approximations}
}
}

\seealso{
\code{\link[=HEQueryCountMaster]{HEQueryCountMaster()}} which goes hand-in-hand with this object
}
\keyword{datasets}
