% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distcomp.R
\name{makeDefinition}
\alias{makeDefinition}
\title{Make a computation definition given the computation type}
\usage{
makeDefinition(compType)
}
\arguments{
\item{compType}{the canonical computation type tag}
}
\value{
a data frame corresponding to the computation type
}
\description{
The function \code{makeDefinition} returns a computational
definition based on current inputs (from the global store) given a
canonical computation type tag. This is a utility function for web
applications to use as input is being gathered
}
\examples{
\dontrun{
makeDefinition(names(availableComputations())[1])
}
}
\seealso{
\code{\link{availableComputations}}
}
