% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distcomp.R
\name{makeHEMaster}
\alias{makeHEMaster}
\title{Instantiate a master process for HE operations}
\usage{
makeHEMaster(defn)
}
\arguments{
\item{defn}{the computation definition}
}
\value{
an master object for HE operations
}
\description{
Instantiate a master process for HE operations
}
