% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/distcomp.R
\name{makeMaster}
\alias{makeMaster}
\title{Make a master object given a definition}
\usage{
makeMaster(defn)
}
\arguments{
\item{defn}{the computation definition}
}
\value{
a master object of the appropriate class based on the definition
}
\description{
The function \code{makeMaster} returns a master object
corresponding to the definition. The types of master objects that can
be created depend upon the available computations
}
\seealso{
\code{\link{availableComputations}}
}

