% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/distcomp.R
\name{makeSlave}
\alias{makeSlave}
\title{Make a slave object given a definition and data}
\usage{
makeSlave(defn, data)
}
\arguments{
\item{defn}{the computation definition}

\item{data}{the data for the computation}
}
\value{
a slave object of the appropriate class based on the definition
}
\description{
The function \code{makeSlave} returns an object of the
appropriate type based on a computation definition and sets the data for
the object. The types of objects that can be created depend upon the
available computations
}
\seealso{
\code{\link{availableComputations}}
}

