% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distcomp.R
\name{makeWorker}
\alias{makeWorker}
\title{Make a worker object given a definition and data}
\usage{
makeWorker(defn, data, pubkey_bits, pubkey_n, den_bits)
}
\arguments{
\item{defn}{the computation definition}

\item{data}{the data for the computation}

\item{pubkey_bits}{the number of bits for the public key}

\item{pubkey_n}{the \code{n} for public key}

\item{den_bits}{the number of bits for the denominator}
}
\value{
a worker object of the appropriate class based on the definition
}
\description{
The function \code{makeWorker} returns an object of the
appropriate type based on a computation definition and sets the data for
the object. The types of objects that can be created depend upon the
available computations
}
\seealso{
\code{\link[=availableComputations]{availableComputations()}}
}
