% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncp.R
\docType{class}
\name{NCP}
\alias{NCP}
\title{Create a non-cooperating object for use in a distributed homomorphic computation}
\format{An \code{\link[R6:R6Class]{R6::R6Class()}} generator object}
\usage{
NCP
}
\description{
\code{NCP} objects are worker objects that separate a
master process from communicating directly with the worker
processes. Typically two such are needed for a distributed
homomorphic computation. A master process can communicate with
\code{NCP} objects and the \code{NCP} objects can communicate
with worker processes. However, the two \code{NCP} objects,
designated by numbers 1 and 2, are non-cooperating in the sense
that they don't communicate with each other and are isolated
from each other.
}
\section{Methods}{

\describe{
\item{\code{NCP$new(name, defn)}}{Create a new \code{NCP} instance with a name and a computation definition}
\item{\code{setParams(pubkey, den)}}{Set a public key, and a denominator for rational approximations of fractional parts of real numbers}
\item{\code{addSite(site)}}{Register a site for a computation}
}
}

\keyword{datasets}
