% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distcomp.R
\name{resetComputationInfo}
\alias{resetComputationInfo}
\title{Clear the contents of the global store}
\usage{
resetComputationInfo()
}
\value{
an empty list
}
\description{
In distcomp, several web applications need to communicate
between themselves. Since only one application is expected to be
active at any time, they do so via a global store, essentially a hash table.
This function clears the store, except for the working directory.
}
\seealso{
\code{\link{setComputationInfo}} \code{\link{getComputationInfo}}
}
