% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/distcomp.R
\name{runDistcompApp}
\alias{runDistcompApp}
\title{Run a specified distcomp web application}
\usage{
runDistcompApp(appType = c("definition", "setupSlave", "setupMaster"))
}
\arguments{
\item{appType}{one of three values: "definition", "setupSlave", "setupMaster"}
}
\value{
the results of running the web application
}
\description{
Web applications can define computation, setup slave sites or masters.
This function invokes the appropriate web application depending on the task
}
\seealso{
\code{\link{defineNewComputation}}, \code{\link{setupSlave}},
\code{\link{setupMaster}}
}

