% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/distcomp.R
\name{setComputationInfo}
\alias{setComputationInfo}
\title{Set a name to a value in a global variable}
\usage{
setComputationInfo(name, value)
}
\arguments{
\item{name}{the name for the object}

\item{value}{the value for the object}
}
\value{
invisibly returns the all the name value pairs
}
\description{
In distcomp, several web applications need to communicate
between themselves. Since only one application is expected to be
active at any time, they do so via a global store, essentially a hash table.
This function sets a name to a value
}
\seealso{
\code{\link{getComputationInfo}}
}

