% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/distcomp.R
\name{setupMaster}
\alias{setupMaster}
\title{Setup a computation master}
\usage{
setupMaster()
}
\value{
the results of running the web application
}
\description{
This function just calls \code{\link{runDistcompApp}} with the
parameter "setupMaster"
}
\seealso{
\code{\link{runDistcompApp}}
}

