% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/distcomp.R
\name{setupSlave}
\alias{setupSlave}
\title{Setup a slave site}
\usage{
setupSlave()
}
\value{
the results of running the web application
}
\description{
This function just calls \code{\link{runDistcompApp}} with the
parameter "setupSlave"
}
\seealso{
\code{\link{runDistcompApp}}
}

