% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/distcomp.R
\name{setupWorker}
\alias{setupWorker}
\title{Setup a worker site}
\usage{
setupWorker()
}
\value{
the results of running the web application
}
\description{
This function just calls \code{\link{runDistcompApp}} with the
parameter "setupWorker"
}
\seealso{
\code{\link{runDistcompApp}}
}

