\name{svdSlave}
\alias{svdSlave}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
An SVD slave site generator function
}
\description{
An SVD slave site generator function
}
\usage{
svdSlave(...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{\dots}{
    Usual arguments for a reference class
}
}
\details{
  Creates a new SVD slave site
}
\value{
A SVD slave site object
}
