% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svdstuff.R
\docType{class}
\name{SVDWorker}
\alias{SVDWorker}
\title{Create a worker object for use as a worker with master objects generated by \code{\link[=SVDMaster]{SVDMaster()}}}
\format{An \code{\link[R6:R6Class]{R6::R6Class()}} generator object}
\description{
\code{SVDWorker} objects are worker objects at each site of
a distributed SVD model computation
}
\section{Methods}{


\describe{
\item{\code{SVDWorker$new(defn, data, stateful = TRUE)}}{Create a new SVD worker object with given definition defn, data and flag for preserving state between iterations}
\item{\code{reset()}}{Initialize work matrix and set up starting values for iterating}
\item{\code{dimX(...)}}{Return the dimensions of the matrix}
\item{\code{updateV(arg, ...)}}{Return an updated value for the V vector, normalized by \code{arg}}
\item{\code{updateU(arg, ...)}}{Return an updated value for the norm of the U vector}
\item{\code{normU(arg, ...)}}{Normalize U vector by \code{arg}}
\item{\code{fixU(arg, ...)}}{Construct the residual matrix using \code{arg}}
\item{\code{getN(...)}}{Return the number of rows}
\item{\code{getP(...)}}{Return the number of columns}
\item{\code{kosher()}}{Check if inputs and state of object are sane. For future use}
\item{\code{getStateful()}}{Returns \code{TRUE} if object is stateful, else \code{FALSE}}
}
}

\seealso{
\code{\link[=SVDMaster]{SVDMaster()}} which goes hand-in-hand with this object
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SVDWorker$new()}}
\item \href{#method-reset}{\code{SVDWorker$reset()}}
\item \href{#method-dimX}{\code{SVDWorker$dimX()}}
\item \href{#method-updateV}{\code{SVDWorker$updateV()}}
\item \href{#method-updateU}{\code{SVDWorker$updateU()}}
\item \href{#method-normU}{\code{SVDWorker$normU()}}
\item \href{#method-fixU}{\code{SVDWorker$fixU()}}
\item \href{#method-getN}{\code{SVDWorker$getN()}}
\item \href{#method-getP}{\code{SVDWorker$getP()}}
\item \href{#method-getStateful}{\code{SVDWorker$getStateful()}}
\item \href{#method-kosher}{\code{SVDWorker$kosher()}}
\item \href{#method-clone}{\code{SVDWorker$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SVDWorker$new(defn, data, stateful = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reset"></a>}}
\subsection{Method \code{reset()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SVDWorker$reset()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dimX"></a>}}
\subsection{Method \code{dimX()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SVDWorker$dimX(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-updateV"></a>}}
\subsection{Method \code{updateV()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SVDWorker$updateV(arg, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-updateU"></a>}}
\subsection{Method \code{updateU()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SVDWorker$updateU(arg, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-normU"></a>}}
\subsection{Method \code{normU()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SVDWorker$normU(arg, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fixU"></a>}}
\subsection{Method \code{fixU()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SVDWorker$fixU(arg, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getN"></a>}}
\subsection{Method \code{getN()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SVDWorker$getN()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getP"></a>}}
\subsection{Method \code{getP()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SVDWorker$getP()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getStateful"></a>}}
\subsection{Method \code{getStateful()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SVDWorker$getStateful()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-kosher"></a>}}
\subsection{Method \code{kosher()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SVDWorker$kosher()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SVDWorker$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
