\name{coxMaster}
\alias{coxMaster}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
An analysis master site generator reference class
}
\description{
An analysis master site generator reference class
}
\usage{
coxMaster(...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{\dots}{
    Usual reference class arguments
}
}
\details{
Creates an analysis master object for computation
}
\value{
  An analysis master object
}
