\name{coxSlave}
\alias{coxSlave}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
A slave site generator function
}
\description{
A slave site generator function
}
\usage{
coxSlave(...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{\dots}{
    Usual arguments for a reference class
}
}
\details{
  Creates a new slave site
}
\value{
A slave site object
}
