% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxstuff.R
\docType{class}
\name{CoxWorker}
\alias{CoxWorker}
\title{Create a worker object for use as a worker with master objects generated by \code{\link[=CoxMaster]{CoxMaster()}}}
\format{An \code{\link[R6:R6Class]{R6::R6Class()}} generator object}
\description{
\code{CoxWorker} objects are worker objects at each site of
a distributed Cox model computation
}
\section{Methods}{

\describe{
\item{\code{CoxWorker$new(defn, data, stateful=TRUE)}}{Create a new CoxWorker instance
object using formula and data. The stateful flag indicates whether the object
state is to be saved between iterations}
\item{\code{logLik(beta, ...)}}{Compute the partial log likelihood for the local data
for the input parameter vector beta. The return value is a named list with
three components: \code{value} contains the value of the log likelihood,
\code{gradient} contains the score vector, and \code{hessian} contains
the estimated hessian matrix}
\item{\code{var(beta, ...)}}{Compute the variance of the parameter vector \code{beta}}
\item{\code{kosher()}}{Check if inputs and state of object are sane. For future use}
\item{\code{getP()}}{Returns the dimension of the parameter \code{vector}}
\item{\code{getStateful()}}{Returns \code{TRUE} if object is stateful, else \code{FALSE}}
}
}

\seealso{
\code{CoxMaster} which goes hand-in-hand with this object
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CoxWorker$new()}}
\item \href{#method-getP}{\code{CoxWorker$getP()}}
\item \href{#method-getStateful}{\code{CoxWorker$getStateful()}}
\item \href{#method-logLik}{\code{CoxWorker$logLik()}}
\item \href{#method-var}{\code{CoxWorker$var()}}
\item \href{#method-kosher}{\code{CoxWorker$kosher()}}
\item \href{#method-clone}{\code{CoxWorker$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoxWorker$new(defn, data, stateful = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getP"></a>}}
\subsection{Method \code{getP()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoxWorker$getP(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getStateful"></a>}}
\subsection{Method \code{getStateful()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoxWorker$getStateful()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-logLik"></a>}}
\subsection{Method \code{logLik()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoxWorker$logLik(beta, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-var"></a>}}
\subsection{Method \code{var()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoxWorker$var(beta, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-kosher"></a>}}
\subsection{Method \code{kosher()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoxWorker$kosher()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoxWorker$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
