% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distcomp.R
\name{destroyInstanceObject}
\alias{destroyInstanceObject}
\title{Destroy an instance object given its identifier}
\usage{
destroyInstanceObject(instanceId)
}
\arguments{
\item{instanceId}{the id of the object to destroy}
}
\value{
TRUE if everything goes well
}
\description{
The function \code{destroyInstanceObject} deletes an object associated
with the instanceId. This is typically done after a computation completes and results
have been obtained.
}
\seealso{
\code{\link[=createInstanceObject]{createInstanceObject()}}
}
