\name{distcomp-internal}
\alias{distcomp-internal}
\alias{.makeOpencpuURL}
\title{Internal distcomp functions}
\description{Internal distcomp functions}
\usage{
.makeOpencpuURL(fn, urlPrefix, package="distcomp")
}
\details{The arguments to these routines are not guarranteed to stay the
  same from release to release -- call them at your own risk!}
\keyword{internal}
