% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distcomp.R
\name{executeMethod}
\alias{executeMethod}
\title{Given the id of a serialized object, invoke a method on the object with arguments}
\usage{
executeMethod(objectId, method, ...)
}
\arguments{
\item{objectId}{the (instance) identifier of the object on which to invoke a method}

\item{method}{the name of the method to invoke}

\item{...}{further arguments as appropriate for the method}
}
\value{
a result that depends on the computation being executed
}
\description{
The function \code{executeMethod} is really the heart of distcomp.
It executes an arbitrary method on an object that has been serialized to the
distcomp workspace with any specified arguments. The result, which is dependent
on the computation that is executed, is returned. If the object needs to save
state between iterations on it, it is automatically serialized back for the ensuing
iterations
}
