% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master.R
\docType{class}
\name{HEMaster}
\alias{HEMaster}
\title{Create a HEMaster process for use in a distributed homomorphic
computation}
\format{An \code{\link[R6:R6Class]{R6::R6Class()}} generator object}
\description{
\code{HEMaster} objects run a distributed computation based
upon a definition file that encapsulates all information
necessary to perform a computation. A master makes use of two
non-cooperating parties which communicate with sites that
perform the actual computations using local data.
}
\section{Methods}{

\describe{
\item{\code{HEMaster$new(defn)}}{Create a new \code{HEMaster} instance using
definition provided}
\item{\code{addNCP(url, ncpWorker)}}{Add a non-cooperating party using either the \code{url} provided or a \code{ncpWorker} object of class \code{NCP}. The latter is for prototyping locally using created objects and so exactly one of the two has to be non-null}
\item{\code{run()}}{Run a computation}
}
}

\seealso{
\code{\link[=NCP]{NCP()}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{HEMaster$new()}}
\item \href{#method-getNC_party}{\code{HEMaster$getNC_party()}}
\item \href{#method-getPubkey}{\code{HEMaster$getPubkey()}}
\item \href{#method-getPrivKey}{\code{HEMaster$getPrivKey()}}
\item \href{#method-addNCP}{\code{HEMaster$addNCP()}}
\item \href{#method-run}{\code{HEMaster$run()}}
\item \href{#method-clone}{\code{HEMaster$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HEMaster$new(defn)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getNC_party"></a>}}
\subsection{Method \code{getNC_party()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HEMaster$getNC_party()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getPubkey"></a>}}
\subsection{Method \code{getPubkey()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HEMaster$getPubkey()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getPrivKey"></a>}}
\subsection{Method \code{getPrivKey()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HEMaster$getPrivKey()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-addNCP"></a>}}
\subsection{Method \code{addNCP()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HEMaster$addNCP(ncp_defn, url = NULL, ncpWorker = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run"></a>}}
\subsection{Method \code{run()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HEMaster$run(debug = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HEMaster$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
