% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/querycount.R
\docType{class}
\name{HEQueryCountMaster}
\alias{HEQueryCountMaster}
\title{Create a master object to control worker objects generated by \code{\link[=HEQueryCountWorker]{HEQueryCountWorker()}}}
\format{An \code{\link[R6:R6Class]{R6::R6Class()}} generator object}
\usage{
HEQueryCountMaster
}
\description{
\code{HEQueryCountMaster} objects instantiate and run a distributed homomorphic QueryCount computation
}
\section{Methods}{


\describe{
\item{\code{HEQueryCountMaster$new(defn, debug = FALSE)}}{Create an HEQueryCount master object
with the specified computation definition. The debug flag is used for debugging computations}
\item{\code{run()}}{Run the QueryCount computation}
}
}

\seealso{
\code{\link[=HEQueryCountWorker]{HEQueryCountWorker()}} which goes hand-in-hand with this object
}
\keyword{datasets}
