% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/querycount.R
\docType{class}
\name{HEQueryCountMaster}
\alias{HEQueryCountMaster}
\title{Create a master object to control worker objects generated by \code{\link[=HEQueryCountWorker]{HEQueryCountWorker()}}}
\format{An \code{\link[R6:R6Class]{R6::R6Class()}} generator object}
\description{
\code{HEQueryCountMaster} objects instantiate and run a distributed homomorphic QueryCount computation; they're instantiated by non-cooperating parties (NCPs)
}
\section{Methods}{


\describe{
\item{\code{HEQueryCountMaster$new(defn, debug = FALSE)}}{Create an HEQueryCount master object
with the specified computation definition. The debug flag is used for debugging computations}
\item{\code{run()}}{Run the QueryCount computation}
}
}

\seealso{
\code{\link[=HEQueryCountWorker]{HEQueryCountWorker()}} which goes hand-in-hand with this object
}
\section{Super class}{
\code{\link[distcomp:QueryCountMaster]{distcomp::QueryCountMaster}} -> \code{HEQueryCountMaster}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{HEQueryCountMaster$new()}}
\item \href{#method-setParams}{\code{HEQueryCountMaster$setParams()}}
\item \href{#method-kosher}{\code{HEQueryCountMaster$kosher()}}
\item \href{#method-queryCount}{\code{HEQueryCountMaster$queryCount()}}
\item \href{#method-cleanup}{\code{HEQueryCountMaster$cleanup()}}
\item \href{#method-run}{\code{HEQueryCountMaster$run()}}
\item \href{#method-summary}{\code{HEQueryCountMaster$summary()}}
\item \href{#method-clone}{\code{HEQueryCountMaster$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="distcomp" data-topic="QueryCountMaster" data-id="addSite">}\href{../../distcomp/html/QueryCountMaster.html#method-addSite}{\code{distcomp::QueryCountMaster$addSite()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="distcomp" data-topic="QueryCountMaster" data-id="getSites">}\href{../../distcomp/html/QueryCountMaster.html#method-getSites}{\code{distcomp::QueryCountMaster$getSites()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HEQueryCountMaster$new(defn, partyNumber, debug = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setParams"></a>}}
\subsection{Method \code{setParams()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HEQueryCountMaster$setParams(pubkey_bits, pubkey_n, den_bits)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-kosher"></a>}}
\subsection{Method \code{kosher()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HEQueryCountMaster$kosher()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-queryCount"></a>}}
\subsection{Method \code{queryCount()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HEQueryCountMaster$queryCount(token)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cleanup"></a>}}
\subsection{Method \code{cleanup()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HEQueryCountMaster$cleanup()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run"></a>}}
\subsection{Method \code{run()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HEQueryCountMaster$run(token)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-summary"></a>}}
\subsection{Method \code{summary()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HEQueryCountMaster$summary()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HEQueryCountMaster$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
