% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/querycount.R
\docType{class}
\name{HEQueryCountWorker}
\alias{HEQueryCountWorker}
\title{Create a worker object for use as a worker with master objects generated by \code{\link[=HEQueryCountMaster]{HEQueryCountMaster()}}}
\format{An \code{\link[R6:R6Class]{R6::R6Class()}} generator object}
\description{
\code{HEQueryCountWorker} objects are worker objects at each site of
a distributed QueryCount model computation using homomorphic encryption
}
\section{Methods}{


\describe{
\item{\code{HEQueryCountWorker$new(defn, data, pubkey, den)}}{Create a new HEQueryCount worker object with given
definition defn, data, flag for preserving state between iterations, public key and denominator for rational approximations}
\item{\code{setParams(pubkey, den)}}{Set the public key and denominator for rational approximations}
}
}

\seealso{
\code{\link[=HEQueryCountMaster]{HEQueryCountMaster()}} which goes hand-in-hand with this object
}
\section{Super class}{
\code{\link[distcomp:QueryCountWorker]{distcomp::QueryCountWorker}} -> \code{HEQueryCountWorker}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{HEQueryCountWorker$new()}}
\item \href{#method-setParams}{\code{HEQueryCountWorker$setParams()}}
\item \href{#method-queryCount}{\code{HEQueryCountWorker$queryCount()}}
\item \href{#method-clone}{\code{HEQueryCountWorker$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="distcomp" data-topic="QueryCountWorker" data-id="getStateful">}\href{../../distcomp/html/QueryCountWorker.html#method-getStateful}{\code{distcomp::QueryCountWorker$getStateful()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="distcomp" data-topic="QueryCountWorker" data-id="kosher">}\href{../../distcomp/html/QueryCountWorker.html#method-kosher}{\code{distcomp::QueryCountWorker$kosher()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HEQueryCountWorker$new(
  defn,
  data,
  pubkey_bits = NULL,
  pubkey_n = NULL,
  den_bits = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setParams"></a>}}
\subsection{Method \code{setParams()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HEQueryCountWorker$setParams(pubkey_bits, pubkey_n, den_bits)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-queryCount"></a>}}
\subsection{Method \code{queryCount()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HEQueryCountWorker$queryCount(partyNumber, token)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HEQueryCountWorker$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
