\name{makeMaster}
\alias{makeMaster}
\title{
Creates a Master object based on the argument defn
}
\description{
Creates a Master object based on the argument defn
}
\usage{
makeMaster(defn)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{defn}{
    Defn is a list consisting of elements that specify the computation
    type among other things.
}
}
\details{
  Creates an Master object
}
\value{
A Master object
}
