% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distcomp.R
\name{makeNCP}
\alias{makeNCP}
\title{Instantiate an noncooperating party}
\usage{
makeNCP(
  defn,
  partyNumber,
  sites = list(),
  pubkey_bits = NULL,
  pubkey_n = NULL,
  den_bits = NULL
)
}
\arguments{
\item{defn}{the computation definition}

\item{partyNumber}{the party number}

\item{sites}{a list of sites each entry a named list of name, url, worker}

\item{pubkey_bits}{number of bits for public key}

\item{pubkey_n}{the n for the public key}

\item{den_bits}{the log to base 2 of the denominator}
}
\value{
an NCP object
}
\description{
Instantiate an noncooperating party
}
