\name{makeSlave}
\alias{makeSlave}
\title{
  Creates a slave object based on the argument bundle
}
\description{
  Creates a slave object based on the argument defn and data
}
\usage{
makeSlave(defn, data)
}

\arguments{
  \item{defn}{
    Defn is a list consisting of elements that specify the computation
    type among other things.
  }
  \item{data}{
    data is a list object containing all pertinent data for the
    computation
  }
}
\details{
  Creates a slave object
}
\value{
  A slave object
}
