% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distcomp.R
\name{makeWorker}
\alias{makeWorker}
\title{Make a worker object given a definition and data}
\usage{
makeWorker(defn, data)
}
\arguments{
\item{defn}{the computation definition}

\item{data}{the data for the computation}
}
\value{
a worker object of the appropriate class based on the definition
}
\description{
The function \code{makeWorker} returns an object of the
appropriate type based on a computation definition and sets the data for
the object. The types of objects that can be created depend upon the
available computations
}
\seealso{
\code{\link{availableComputations}}
}
