% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/querycount.R
\docType{class}
\name{QueryCountMaster}
\alias{QueryCountMaster}
\title{Create a master object to control worker objects generated by \code{\link[=QueryCountWorker]{QueryCountWorker()}}}
\format{An \code{\link[R6:R6Class]{R6::R6Class()}} generator object}
\description{
\code{QueryCountMaster} objects instantiate and run a distributed QueryCount computation
}
\section{Methods}{


\describe{
\item{\code{QueryCountMaster$new(defn, debug = FALSE)}}{Create an QueryCount master object
with the specified computation definition. The debug flag is used for debugging computations}
\item{\code{kosher()}}{Check if inputs and state of object are sane. For future use}
\item{\code{run()}}{Run the QueryCount computation until either the
threshold is reached or the max.iter number of iterations are used up}
\item{\code{summary()}}{Return the summary of results}
}
}

\seealso{
\code{\link[=QueryCountWorker]{QueryCountWorker()}} which goes hand-in-hand with this object
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{QueryCountMaster$new()}}
\item \href{#method-kosher}{\code{QueryCountMaster$kosher()}}
\item \href{#method-queryCount}{\code{QueryCountMaster$queryCount()}}
\item \href{#method-getSites}{\code{QueryCountMaster$getSites()}}
\item \href{#method-addSite}{\code{QueryCountMaster$addSite()}}
\item \href{#method-run}{\code{QueryCountMaster$run()}}
\item \href{#method-summary}{\code{QueryCountMaster$summary()}}
\item \href{#method-clone}{\code{QueryCountMaster$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QueryCountMaster$new(defn, debug = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-kosher"></a>}}
\subsection{Method \code{kosher()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QueryCountMaster$kosher()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-queryCount"></a>}}
\subsection{Method \code{queryCount()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QueryCountMaster$queryCount()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getSites"></a>}}
\subsection{Method \code{getSites()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QueryCountMaster$getSites()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-addSite"></a>}}
\subsection{Method \code{addSite()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QueryCountMaster$addSite(name, url = NULL, worker = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run"></a>}}
\subsection{Method \code{run()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QueryCountMaster$run()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-summary"></a>}}
\subsection{Method \code{summary()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QueryCountMaster$summary()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QueryCountMaster$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
