% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/querycount.R
\docType{class}
\name{QueryCountWorker}
\alias{QueryCountWorker}
\title{Create a worker object for use as a worker with master objects generated by \code{\link[=QueryCountMaster]{QueryCountMaster()}}}
\format{An \code{\link[R6:R6Class]{R6::R6Class()}} generator object}
\usage{
QueryCountWorker
}
\description{
\code{QueryCountWorker} objects are worker objects at each site of
a distributed QueryCount model computation
}
\section{Methods}{


\describe{
\item{\code{QueryCountWorker$new(defn, data)}}{Create a new QueryCount worker object with given definition defn, data and flag for preserving state between iterations}
\item{\code{kosher()}}{Check if inputs and state of object are sane. For future use}
\item{\code{getStateful()}}{Returns \code{FALSE}}
}
}

\seealso{
\code{\link[=QueryCountMaster]{QueryCountMaster()}} which goes hand-in-hand with this object
}
\keyword{datasets}
