% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/querycount.R
\docType{class}
\name{QueryCountWorker}
\alias{QueryCountWorker}
\title{Create a worker object for use as a worker with master objects generated by \code{\link[=QueryCountMaster]{QueryCountMaster()}}}
\format{An \code{\link[R6:R6Class]{R6::R6Class()}} generator object}
\description{
\code{QueryCountWorker} objects are worker objects at each site of
a distributed QueryCount model computation
}
\section{Methods}{


\describe{
\item{\code{QueryCountWorker$new(defn, data)}}{Create a new QueryCount worker object with given definition defn, data and flag for preserving state between iterations}
\item{\code{kosher()}}{Check if inputs and state of object are sane. For future use}
\item{\code{getStateful()}}{Returns \code{FALSE}}
}
}

\seealso{
\code{\link[=QueryCountMaster]{QueryCountMaster()}} which goes hand-in-hand with this object
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{QueryCountWorker$new()}}
\item \href{#method-getStateful}{\code{QueryCountWorker$getStateful()}}
\item \href{#method-kosher}{\code{QueryCountWorker$kosher()}}
\item \href{#method-queryCount}{\code{QueryCountWorker$queryCount()}}
\item \href{#method-clone}{\code{QueryCountWorker$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QueryCountWorker$new(defn, data, stateful = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getStateful"></a>}}
\subsection{Method \code{getStateful()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QueryCountWorker$getStateful()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-kosher"></a>}}
\subsection{Method \code{kosher()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QueryCountWorker$kosher()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-queryCount"></a>}}
\subsection{Method \code{queryCount()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QueryCountWorker$queryCount()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QueryCountWorker$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
