% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distcomp.R
\name{runDistcompApp}
\alias{runDistcompApp}
\title{Run a specified distcomp web application}
\usage{
runDistcompApp(appType = c("definition", "setupWorker", "setupMaster"))
}
\arguments{
\item{appType}{one of three values: "definition", "setupWorker", "setupMaster"}
}
\value{
the results of running the web application
}
\description{
Web applications can define computation, setup worker sites or masters.
This function invokes the appropriate web application depending on the task
}
\seealso{
\code{\link[=defineNewComputation]{defineNewComputation()}}, \code{\link[=setupWorker]{setupWorker()}}, \code{\link[=setupMaster]{setupMaster()}}
}
