\name{svdMaster}
\alias{svdMaster}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
An SVD analysis master site generator reference class
}
\description{
An SVD analysis master site generator reference class
}
\usage{
svdMaster(...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{\dots}{
    Usual reference class arguments
}
}
\details{
Creates an analysis master object for SVD computation
}
\value{
  An SVD analysis master object
}
