% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svdstuff.R
\docType{class}
\name{SVDMaster}
\alias{SVDMaster}
\title{Create a master object to control worker objects generated by \code{\link[=SVDWorker]{SVDWorker()}}}
\format{An \code{\link[R6:R6Class]{R6::R6Class()}} generator object}
\description{
\code{SVDMaster} objects instantiate and run a distributed SVD computation
}
\section{Methods}{


\describe{
\item{\code{SVDMaster$new(defn, debug = FALSE)}}{Create an SVD master object
with the specified computation definition. The debug flag is used for debugging computations}
\item{\code{kosher()}}{Check if inputs and state of object are sane. For future use}
\item{\code{updateV(arg)}}{Return an updated value for the V vector}
\item{\code{updateU(arg)}}{Return an updated value for the U vector}
\item{\code{fixFit(v, d)}}{Construct the residual matrix using given the v vector and d so far}
\item{\code{reset()}}{Initialize the computation}
\item{\code{dimX(, ...)}}{Return the dimensions of the matrix}
\item{\code{normU(arg, ...)}}{Normalize U vector by \code{arg}}
\item{\code{addSite(name, url = NULL, worker = NULL)}}{Add a worker site
for participating in the distributed computation. Exactly one of \code{url}
or \code{worker} should be specified}
\item{\code{run(k = private$k, thr = 1e-8, max.iter = 100)}}{Run the SVD computation until either the
threshold is reached or the max.iter number of iterations are used up}
\item{\code{summary()}}{Return the summary of results}
}
}

\seealso{
\code{\link[=SVDWorker]{SVDWorker()}} which goes hand-in-hand with this object
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SVDMaster$new()}}
\item \href{#method-kosher}{\code{SVDMaster$kosher()}}
\item \href{#method-updateV}{\code{SVDMaster$updateV()}}
\item \href{#method-updateU}{\code{SVDMaster$updateU()}}
\item \href{#method-fixFit}{\code{SVDMaster$fixFit()}}
\item \href{#method-reset}{\code{SVDMaster$reset()}}
\item \href{#method-addSite}{\code{SVDMaster$addSite()}}
\item \href{#method-run}{\code{SVDMaster$run()}}
\item \href{#method-summary}{\code{SVDMaster$summary()}}
\item \href{#method-clone}{\code{SVDMaster$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SVDMaster$new(defn, debug = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-kosher"></a>}}
\subsection{Method \code{kosher()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SVDMaster$kosher()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-updateV"></a>}}
\subsection{Method \code{updateV()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SVDMaster$updateV(arg)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-updateU"></a>}}
\subsection{Method \code{updateU()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SVDMaster$updateU(arg)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fixFit"></a>}}
\subsection{Method \code{fixFit()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SVDMaster$fixFit(v, d)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reset"></a>}}
\subsection{Method \code{reset()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SVDMaster$reset()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-addSite"></a>}}
\subsection{Method \code{addSite()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SVDMaster$addSite(name, url = NULL, worker = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run"></a>}}
\subsection{Method \code{run()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SVDMaster$run(thr = 1e-08, max.iter = 100)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-summary"></a>}}
\subsection{Method \code{summary()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SVDMaster$summary()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SVDMaster$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
