% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distcomp.R
\name{uploadNewComputation}
\alias{uploadNewComputation}
\title{Upload a new computation and data to an opencpu server}
\usage{
uploadNewComputation(site, defn, data)
}
\arguments{
\item{site}{a list of two items, a unique \code{name} and a \code{url}}

\item{defn}{the identifier of an already defined computation}

\item{data}{the (local) data to use}
}
\value{
TRUE if everything goes well
}
\description{
The function \code{uploadNewComputation} is really a remote version
of \code{\link{saveNewComputation}}, invoking that function on an opencpu server.
This is typically done for every site that wants to participate in a computation
with its own local data. Note that a site is always a list of at least a unique
name element (distinguishing the site from others) and a url element.
}
\seealso{
\code{\link{saveNewComputation}}
}
